Implement advanced search in `search-service` with Elasticsearch integration:
- Add `SearchRequest` and `PageResponse` DTOs for request and response handling.
- Create `ProductDocumentController` to expose search endpoint.
- Update `ProductDocumentService` to support dynamic filtering using Elasticsearch queries.
- Refactor `price` field to use `BigDecimal` in relevant classes for consistency.
- Enhance `GatewayConfiguration` to route search endpoints to `search-service`.

+++ product-service/src/main/java/com/javabuilder/productservice/service/impl/ProductServiceImpl.java
@@ -67,7 +67,7 @@ public class ProductServiceImpl implements ProductService {
                 .id(product.getId())
                 .name(product.getName())
                 .description(product.getDescription())
-                .price(product.getPrice().doubleValue())
+                .price(product.getPrice())
                 .categoryId(category.getId())
                 .categoryName(category.getName())
                 .thumbnail(product.getImages().stream().filter(ProductImage::getIsPrimary).map(ProductImage::getUrl).findFirst().orElse(""))
